+++ src/Kiota.Builder/Refiners/ILanguageRefiner.cs
@@ -40,6 +40,9 @@ public interface ILanguageRefiner
             case GenerationLanguage.Python:
                 await new PythonRefiner(config).RefineAsync(generatedCode, cancellationToken).ConfigureAwait(false);
                 break;
+            case GenerationLanguage.Dart:
+                await new DartRefiner(config).RefineAsync(generatedCode, cancellationToken).ConfigureAwait(false);
+                break;
         }
     }
 }