+++ vscode/microsoft-kiota/CHANGELOG.md
@@ -6,7 +6,18 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Added dart generation to Visual Studio Code extension.
+
 ### Changed
+
+- Fixed a bug in the VS Code extension deeplink with the API Center extension [#6004](https://github.com/microsoft/kiota/issues/6004)
+
+## [1.22.100000001] - 2025-01-10
+
+### Added
+
+### Changed
+- Fixed a bug in the VS Code extension plugin generation [#5978](https://github.com/microsoft/kiota/issues/5978)
 
 ## [1.20.000000001] - 2024-11-07